Add quota page and form
Added and cleaned up the quota page before creating the form.
Made it 2 rows of 4 items instead of the initial 3 by 3.
Added border to the items as well.

Removed ContactDetails from Router.
Fixed image sources for some pages.
Added new images for the Home page.
Added 'react-bootstrap' as a new dependency
in order to create the form more easily.

Finally added the quota form page and implemented only
the basic UI that is common for all quota requests.